+++ src/http_client.rs
@@ -2,15 +2,14 @@ use hyper::{
     client::{connect::Connect, ResponseFuture},
     Body, Request,
 };
+use sealed::sealed;
 
-use crate::sealed::Sealed;
-
-pub trait HttpClient: Sealed + Send + Sync + 'static {
+#[sealed]
+pub trait HttpClient: Send + Sync + 'static {
     fn _request(&self, req: Request<Body>) -> ResponseFuture;
 }
 
-impl<C> Sealed for hyper::Client<C> {}
-
+#[sealed]
 impl<C> HttpClient for hyper::Client<C>
 where
     C: Connect + Clone + Send + Sync + 'static,